adding normalization function

+++ Project3/AI Project 3/Assets/Scripts/MapGeneration/GridMap.cs
@@ -16,7 +16,7 @@ public class GridMap{
 	public TileTypes currentTile;
 	public TileTypes[,] gridData { get; set; }
 	public Node[,] graph { get; set; }
-	public double[,] probabilities;
+	public float[,] probabilities;
 
 
 	//Singleton to make sure only 1 map exists at a time
@@ -86,12 +86,12 @@ public class GridMap{
 		x_columns++; y_rows++;
 
 		gridData = new TileTypes[x_columns, y_rows];
-		probabilities = new double[x_columns, y_rows];
+		probabilities = new float[x_columns, y_rows];
 		for (uint i = 1; i < 4; i++)
 		{
 			for (uint j = 1; j < 4; j++)
 			{
-				probabilities[i, j] = 1 / 8.0;
+                probabilities[i, j] = (1f / 8.0f);
 			}
 		}
 	}
